# api/kube-jit/docs/swagger.yaml
basePath: /kube-jit-api
definitions:
...
  handlers.CleanExpiredResponse:
    properties:
      deleted:
        type: integer
      message:
        type: string
    type: object
...
  handlers.CommonPermissionsResponse:
    properties:
      adminGroups:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      approverGroups:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      isAdmin:
        type: boolean
      isApprover:
        type: boolean
      isPlatformApprover:
        type: boolean
      platformApproverGroups:
        items:
          $ref: '#/definitions/models.Team'
        type: array
    type: object
...
  models.SimpleMessageResponse:
    properties:
      error:
        type: string
      message:
        type: string
      status:
        type: string
    type: object
  models.Team:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
  description: Self-service Kubernetes RBAC JIT Requests with Google/Azure/Github
    Oauth.
  title: Kube-JIT API
  version: "1.0"
paths:
...
  /kube-jit-api/admin/clean-expired:
    post:
      consumes:
      - application/json
      description: |-
        Deletes JIT requests where endDate is in the past and status is "Requested" (not Approved or Rejected). Admin only.
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expired non-approved requests cleaned
          schema:
            $ref: '#/definitions/handlers.CleanExpiredResponse'
        "401":
          description: 'Unauthorized: admin only'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Failed to clean expired requests
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Clean up expired non-approved JIT requests
      tags:
      - admin
...
  /kube-jit-api/permissions:
    post:
      consumes:
      - application/json
      description: |-
        Returns the user's permissions and group memberships for the specified provider (GitHub, Google, Azure).
        Requires one or more cookies named kube_jit_session_<number> (e.g., kube_jit_session_0, kube_jit_session_1).
        Pass split cookies in the Cookie header, for example:
        -H "Cookie: kube_jit_session_0=${cookie_0};kube_jit_session_1=${cookie_1}"
        Note: Swagger UI cannot send custom Cookie headers due to browser security restrictions. Use curl for testing with split cookies.
      parameters:
      - description: 'Session cookies (multiple allowed, names: kube_jit_session_0,
          kube_jit_session_1, etc.)'
        in: header
        name: Cookie
        required: true
        type: string
      - description: Provider payload (e.g., {\
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: User permissions and groups
          schema:
            $ref: '#/definitions/handlers.CommonPermissionsResponse'
        "400":
          description: Missing or invalid provider
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "401":
          description: 'Unauthorized: no token in session data'
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
        "500":
          description: Failed to fetch user groups
          schema:
            $ref: '#/definitions/models.SimpleMessageResponse'
      summary: Get common permissions for the logged in user
      tags:
      - auth
...
swagger: "2.0"